requirements.txt
streamlit==1.21.0
tensorflow==2.10.0
pillow==9.0.1
numpy==1.21.6
pandas==1.3.5
openai==0.27.0
gdown==4.4.0
